rename RelatedListWithRecommendList
